requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
mcp==1.1.2
pytest>=7.4.0
pytest-asyncio>=0.21.0
